requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
astro-sdk-python
astro-projects
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
